chore: add rpc validation
