+++ packages/sysweb3-utils/src/contracts.ts
@@ -1,6 +1,7 @@
 import { web3Provider } from '@pollum-io/sysweb3-network';
 import { Contract, ContractOptions } from 'web3-eth-contract';
 import { AbiItem } from 'web3-utils';
+import { ethers } from 'ethers';
 
 export const createContractUsingAbi = async (
   AbiContract: AbiItem[] | AbiItem | object,
@@ -12,4 +13,18 @@ export const createContractUsingAbi = async (
     address,
     options
   );
+};
+
+const InfuraProvider = ethers.providers.InfuraProvider;
+
+export const isContractAddress = async (address: string, chainId = 1) => {
+  if (address) {
+    const provider = new InfuraProvider(
+      chainId,
+      "c42232a29f9d4bd89d53313eb16ec241"
+    );
+    const code = await provider.getCode(address);
+    return code !== "0x";
+  }
+  return false;
 };
