+++ packages/sysweb3-utils/src/index.ts
@@ -12,3 +12,6 @@ export * from './keyring';
 export * from './networks';
 export * from './spsbt';
 export * from './signer-info';
+export * from './address';
+export * from './constants';
+export * from './rpc';
